// MiniHotel.Client/src/app/api/models/user-update-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserRole } from '../models/user-role';
export interface UserUpdateDto {
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  phoneNumber?: string | null;
  role?: UserRole;
}